Move 8-bit integer into primitives
