+++ library/Rattletrap.hs
@@ -18,7 +18,7 @@ module Rattletrap
   , module Rattletrap.Header
   , module Rattletrap.Initialization
   , module Rattletrap.Int32
-  , module Rattletrap.Int8
+  , module Rattletrap.Primitive.Int8
   , module Rattletrap.Int8Vector
   , module Rattletrap.KeyFrame
   , module Rattletrap.List
@@ -60,7 +60,7 @@ import Rattletrap.Frame
 import Rattletrap.Header
 import Rattletrap.Initialization
 import Rattletrap.Int32
-import Rattletrap.Int8
+import Rattletrap.Primitive.Int8
 import Rattletrap.Int8Vector
 import Rattletrap.Json ()
 import Rattletrap.KeyFrame
